+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
 ## v0.1.0
-- ListBuckets
-- GetBucket
+- Get Bucket Acl	获得Bucket访问权限
+- Get Bucket Location	获得Bucket所属的数据中心位置信息
+- Get Bucket Logging	查看Bucket的访问日志配置情况
+- Get Bucket Website	查看Bucket的静态网站托管状态
+- Get Bucket Referer	查看Bucket的防盗链规则
+- Get Bucket Lifecycle	查看Bucket中Object的生命周期规则
+- Get Bucket(List Object)	获得Bucket中所有Object的信息
+- Get Bucket Info	获取Bucket信息